Move latest-job tests to dedicated file
