+++ tests/v2/test_09_history/test_history_api.py
@@ -34,142 +34,6 @@ def _assert_dict_equal_with_timestamps(
             assert value1 == value2
 
 
-async def test_status_api(
-    project_factory,
-    workflow_factory,
-    task_factory,
-    dataset_factory,
-    workflowtask_factory,
-    job_factory,
-    db,
-    client,
-    MockCurrentUser,
-):
-    async with MockCurrentUser() as user:
-        project = await project_factory(user)
-        dataset = await dataset_factory(project_id=project.id)
-        workflow = await workflow_factory(project_id=project.id)
-        task = await task_factory(user_id=user.id)
-
-        # WorkflowTask 1 (one run, four units, different statuses)
-        wftask1 = await workflowtask_factory(
-            workflow_id=workflow.id, task_id=task.id
-        )
-        job = await job_factory(
-            project_id=project.id,
-            dataset_id=dataset.id,
-            workflow_id=workflow.id,
-            working_dir="/foo",
-            status="done",
-        )
-        run1 = HistoryRun(
-            workflowtask_id=wftask1.id,
-            dataset_id=dataset.id,
-            workflowtask_dump={},
-            task_group_dump={},
-            num_available_images=3,
-            status=HistoryUnitStatus.SUBMITTED,
-            job_id=job.id,
-        )
-        db.add(run1)
-        await db.commit()
-        await db.refresh(run1)
-
-        unit_a = HistoryUnit(
-            history_run_id=run1.id,
-            status=HistoryUnitStatus.SUBMITTED,
-            logfile="/log/a",
-            zarr_urls=["/a"],
-        )
-        unit_b = HistoryUnit(
-            history_run_id=run1.id,
-            status=HistoryUnitStatus.DONE,
-            logfile="/log/b",
-            zarr_urls=["/b"],
-        )
-        unit_c = HistoryUnit(
-            history_run_id=run1.id,
-            status=HistoryUnitStatus.FAILED,
-            logfile="/log/c",
-            zarr_urls=["/c"],
-        )
-        db.add(unit_a)
-        db.add(unit_b)
-        db.add(unit_c)
-        await db.commit()
-        await db.refresh(unit_a)
-        await db.refresh(unit_b)
-        await db.refresh(unit_c)
-
-        db.add(
-            HistoryImageCache(
-                zarr_url="/a",
-                workflowtask_id=wftask1.id,
-                dataset_id=dataset.id,
-                latest_history_unit_id=unit_a.id,
-            )
-        )
-        db.add(
-            HistoryImageCache(
-                zarr_url="/b",
-                workflowtask_id=wftask1.id,
-                dataset_id=dataset.id,
-                latest_history_unit_id=unit_b.id,
-            )
-        )
-        db.add(
-            HistoryImageCache(
-                zarr_url="/c",
-                workflowtask_id=wftask1.id,
-                dataset_id=dataset.id,
-                latest_history_unit_id=unit_c.id,
-            )
-        )
-        await db.commit()
-
-        wftask2 = await workflowtask_factory(
-            workflow_id=workflow.id, task_id=task.id
-        )
-
-        res = await client.get(
-            f"/api/v2/project/{project.id}/latest-job/"
-            f"?workflow_id={workflow.id}&dataset_id={dataset.id}"
-        )
-        assert res.status_code == 200
-        debug(res.json())
-        assert res.json()["task_statuses"] == {
-            str(wftask1.id): {
-                "status": "submitted",
-                "num_available_images": 3,
-                "num_done_images": 1,
-                "num_submitted_images": 1,
-                "num_failed_images": 1,
-            },
-            str(wftask2.id): None,
-        }
-
-        # Invalid `num_available_images`
-        run1.num_available_images = 2
-        db.add(run1)
-        await db.commit()
-        db.expunge_all()
-        res = await client.get(
-            f"/api/v2/project/{project.id}/latest-job/"
-            f"?workflow_id={workflow.id}&dataset_id={dataset.id}"
-        )
-        assert res.status_code == 200
-        assert res.json()["task_statuses"] == {
-            "1": {
-                "status": "submitted",
-                "num_available_images": None,
-                "num_submitted_images": 1,
-                "num_done_images": 1,
-                "num_failed_images": 1,
-            },
-            "2": None,
-        }
-
-
 @pytest.mark.parametrize(
     "object_to_delete",
     [
